requirements/base.txt
Flask==2.0.2
flask-marshmallow==0.14.0
redis==3.5.3
flask-jwt-extended==4.3.1
Flask-Security==3.0.0
flask-restx==0.5.1
gevent==21.8.0
python-dotenv==0.19.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
sqlalchemy==1.4.25
flask_sqlalchemy==2.5.1
email-validator==1.1.3
bcrypt==3.2.0
loguru==0.5.3